fix(alipay): RSA2 验签串保留 sign_type + topup 金额支持小数
两处修复,均由真实沙箱联调暴露:

1. controller/alipay_rsa2.go — alipaySignParams 不再过滤 sign_type

   原本按"官方文档"口径排除了 sign_type,导致沙箱网关报
   invalid-signature。沙箱网关返回的 canonical 验签串实际包含
   sign_type=RSA2:
     app_id=...&biz_content=...&charset=utf-8&method=...&
     notify_url=...&return_url=...&sign_type=RSA2&timestamp=...&version=1.0

   修复后本地用应用公钥对网关给出的字符串验签通过(逐字节一致)。
   该口径同时适用于异步通知验签(对应 Alipay rsaCheckV2:
   仅排除 sign,保留 sign_type)。

2. controller/topup_alipay.go — TopUpAlipayPayRequest.Amount int -> float64

   沙箱最小金额 ¥0.01,整数类型无法表达,最低金额校验恒失败。

验证:
  POST /api/user/alipay -> 200,签发 pay_url
  收银台可达: 302 -> excashier-sandbox.dl.alipaydev.com/standard/auth.htm
  本地 RSA2 验签: 通过

+++ controller/topup_alipay.go
@@ -29,8 +29,8 @@ import (
 )
 
 type TopUpAlipayPayRequest struct {
-	Amount    int    `json:"amount"`     // 元
-	TopUpCode string `json:"top_up_code"` // 兼容老 topup 字段
+	Amount    float64 `json:"amount"`     // 元(支持小数,沙箱测试 ¥0.01 必须 float64)
+	TopUpCode string  `json:"top_up_code"` // 兼容老 topup 字段
 }
 
 func RequestAlipayPay(c *gin.Context) {
@@ -61,10 +61,10 @@ func RequestAlipayPay(c *gin.Context) {
 		c.JSON(http.StatusOK, gin.H{"success": false, "message": "参数错误"})
 		return
 	}
-	if req.Amount < int(setting.AlipayMinTopUp) {
+	if req.Amount < setting.AlipayMinTopUp {
 		c.JSON(http.StatusOK, gin.H{
 			"success": false,
-			"message": fmt.Sprintf("最低充值金额 %d 元", int(setting.AlipayMinTopUp)),
+			"message": fmt.Sprintf("最低充值金额 %.2f 元", setting.AlipayMinTopUp),
 		})
 		return
 	}
@@ -78,7 +78,7 @@ func RequestAlipayPay(c *gin.Context) {
 	tradeNo := "TU-ALI-" + randstr.String(8) + "-" + strconv.FormatInt(time.Now().Unix(), 10)
 
 	// amount 元 -> quota (1元 = 500000 quota)
-	quota := int64(req.Amount) * 500000
+	quota := int64(req.Amount * 500000)
 
 	topUp := &model.TopUp{
 		UserId:        id,
